+++ common/src/main/java/com/jship/hauntfurnace/menu/HauntFurnaceMenu.java
@@ -2,13 +2,17 @@ package com.jship.hauntfurnace.menu;
 
 import com.jship.hauntfurnace.HauntFurnace.ModMenus;
 import com.jship.hauntfurnace.HauntFurnace.ModRecipes;
+import com.jship.hauntfurnace.block.entity.HauntFurnaceBlockEntity;
 
+import lombok.extern.slf4j.Slf4j;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractFurnaceMenu;
 import net.minecraft.world.inventory.ContainerData;
 import net.minecraft.world.inventory.RecipeBookType;
+import net.minecraft.world.item.ItemStack;
 
+@Slf4j
 public class HauntFurnaceMenu extends AbstractFurnaceMenu {
 
     public HauntFurnaceMenu(int containerId, Inventory playerInventory) {
@@ -32,5 +36,10 @@ public class HauntFurnaceMenu extends AbstractFurnaceMenu {
                 playerInventory,
                 container,
                 containerData);
+    }
+
+    @Override
+    protected boolean isFuel(ItemStack stack) {
+        return HauntFurnaceBlockEntity.isFuel(this.level.fuelValues(), stack);
     }
 }