+++ fabric/src/main/java/com/jship/hauntfurnace/network/fabric/Payloads.java
@@ -13,7 +13,7 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.world.item.Item;
 
 public class Payloads {
-    public static StreamCodec<RegistryFriendlyByteBuf, Map<Item, Integer>> STREAM_CODEC = ByteBufCodecs.map(
+    public static StreamCodec<RegistryFriendlyByteBuf, Map<Item, Integer>> FUEL_MAP_STREAM_CODEC = ByteBufCodecs.map(
             HashMap::new,
             ByteBufCodecs.fromCodec(BuiltInRegistries.ITEM.byNameCodec()),
             ByteBufCodecs.INT,
@@ -21,7 +21,7 @@ public class Payloads {
 
     public record HauntFurnaceFuelMapS2CPayload(Map<Item, Integer> fuelMap) implements CustomPacketPayload {
         public static final CustomPacketPayload.Type<HauntFurnaceFuelMapS2CPayload> ID = new CustomPacketPayload.Type<>(HauntFurnace.id("haunt_furnace_fuel_map"));
-        public static final StreamCodec<RegistryFriendlyByteBuf, HauntFurnaceFuelMapS2CPayload> CODEC = StreamCodec.composite(STREAM_CODEC, HauntFurnaceFuelMapS2CPayload::fuelMap, HauntFurnaceFuelMapS2CPayload::new);
+        public static final StreamCodec<RegistryFriendlyByteBuf, HauntFurnaceFuelMapS2CPayload> CODEC = StreamCodec.composite(FUEL_MAP_STREAM_CODEC, HauntFurnaceFuelMapS2CPayload::fuelMap, HauntFurnaceFuelMapS2CPayload::new);
 
         @Override
         public Type<? extends CustomPacketPayload> type() {
@@ -31,7 +31,7 @@ public class Payloads {
 
     public record EnderFurnaceFuelMapS2CPayload(Map<Item, Integer> fuelMap) implements CustomPacketPayload {
         public static final CustomPacketPayload.Type<EnderFurnaceFuelMapS2CPayload> ID = new CustomPacketPayload.Type<>(HauntFurnace.id("ender_furnace_fuel_map"));
-        public static final StreamCodec<RegistryFriendlyByteBuf, EnderFurnaceFuelMapS2CPayload> CODEC = StreamCodec.composite(STREAM_CODEC, EnderFurnaceFuelMapS2CPayload::fuelMap, EnderFurnaceFuelMapS2CPayload::new);
+        public static final StreamCodec<RegistryFriendlyByteBuf, EnderFurnaceFuelMapS2CPayload> CODEC = StreamCodec.composite(FUEL_MAP_STREAM_CODEC, EnderFurnaceFuelMapS2CPayload::fuelMap, EnderFurnaceFuelMapS2CPayload::new);
 
         @Override
         public Type<? extends CustomPacketPayload> type() {