+++ neoforge/src/main/java/com/jship/hauntfurnace/neoforge/HauntFurnaceNeoforge.java
@@ -1,21 +1,28 @@
-package com.jship.hauntfurnace;
+package com.jship.hauntfurnace.neoforge;
 
+import com.jship.hauntfurnace.HauntFurnace;
 import com.jship.hauntfurnace.HauntFurnace.ModBlockEntities;
 import com.jship.hauntfurnace.HauntFurnace.ModBlocks;
 import com.jship.hauntfurnace.HauntFurnace.ModRecipes;
 import com.jship.hauntfurnace.block.entity.PoweredEnderFurnaceBlockEntity;
 import com.jship.hauntfurnace.block.entity.PoweredHauntFurnaceBlockEntity;
+import com.jship.hauntfurnace.config.HauntFurnaceConfig;
+import com.jship.hauntfurnace.data.neoforge.FuelData;
 import com.jship.hauntfurnace.energy.neoforge.EnergyStorageFactoryNeoforge;
 
 import net.minecraft.world.item.CreativeModeTabs;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLLoadCompleteEvent;
 import net.neoforged.neoforge.capabilities.Capabilities;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
 import net.neoforged.neoforge.client.event.RegisterRecipeBookSearchCategoriesEvent;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import net.neoforged.neoforge.energy.EnergyStorage;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 import net.neoforged.neoforge.items.wrapper.SidedInvWrapper;
+import net.neoforged.neoforge.registries.datamaps.RegisterDataMapTypesEvent;
 
 @Mod(HauntFurnace.MOD_ID)
 public class HauntFurnaceNeoforge {
@@ -25,8 +32,17 @@ public class HauntFurnaceNeoforge {
         HauntFurnace.ENERGY_STORAGE_FACTORY = () -> new EnergyStorageFactoryNeoforge();
 
         modEventBus.addListener(this::addCreative);
+        modEventBus.addListener(this::loadComplete);
         modEventBus.addListener(this::registerCapabilities);
-        modEventBus.addListener(this::registerSearchCategories);
+        modEventBus.addListener(this::registerFuelMaps);
+
+        ModLoadingContext.get().registerExtensionPoint(
+                IConfigScreenFactory.class,
+                () -> (client, parent) -> HauntFurnaceConfig.createConfig(parent));
+    }
+
+    private void loadComplete(final FMLLoadCompleteEvent event) {
+        HauntFurnace.ENERGY_STORAGE_FACTORY = EnergyStorageFactoryNeoforge::new;
     }
 
     private void addCreative(BuildCreativeModeTabContentsEvent event) {
@@ -46,7 +62,9 @@ public class HauntFurnaceNeoforge {
                 Capabilities.EnergyStorage.BLOCK,
                 ModBlockEntities.POWERED_HAUNT_FURNACE.get(),
                 (blockEntity, context) -> (EnergyStorage) ((PoweredHauntFurnaceBlockEntity) blockEntity).energyStorage);
-        event.registerBlockEntity(Capabilities.ItemHandler.BLOCK, ModBlockEntities.HAUNT_FURNACE.get(),
+        event.registerBlockEntity(
+                Capabilities.ItemHandler.BLOCK,
+                ModBlockEntities.HAUNT_FURNACE.get(),
                 (blockEntity, side) -> {
                     return new SidedInvWrapper(blockEntity, side);
                 });
@@ -56,11 +74,14 @@ public class HauntFurnaceNeoforge {
                 (blockEntity, side) -> {
                     return new SidedInvWrapper(blockEntity, side);
                 });
+
         event.registerBlockEntity(
                 Capabilities.EnergyStorage.BLOCK,
                 ModBlockEntities.POWERED_ENDER_FURNACE.get(),
                 (blockEntity, context) -> (EnergyStorage) ((PoweredEnderFurnaceBlockEntity) blockEntity).energyStorage);
-        event.registerBlockEntity(Capabilities.ItemHandler.BLOCK, ModBlockEntities.ENDER_FURNACE.get(),
+        event.registerBlockEntity(
+                Capabilities.ItemHandler.BLOCK,
+                ModBlockEntities.ENDER_FURNACE.get(),
                 (blockEntity, side) -> {
                     return new SidedInvWrapper(blockEntity, side);
                 });
@@ -82,5 +103,10 @@ public class HauntFurnaceNeoforge {
                 ModRecipes.CORRUPTING_BLOCKS_CATEGORY.get(),
                 ModRecipes.CORRUPTING_FOOD_CATEGORY.get(),
                 ModRecipes.CORRUPTING_MISC_CATEGORY.get());
+    }
+
+    private void registerFuelMaps(RegisterDataMapTypesEvent event) {
+        event.register(FuelData.HAUNT_FUEL_MAP);
+        event.register(FuelData.ENDER_FUEL_MAP);
     }
 }